+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Allow to exempt favorites from "Move to top on paste"
 - Preview popup now shows the items's description (if set)
 - Favorites dialog text fields (text, description) now support usual copy/select actions (Cmd+V/A/X/C/Z/Y)
+- Favorites tabs: right-click any entry to insert a horizontal separator below it; right-click the separator to remove it
 
 ## [1.0.1] - 2026-04-02
 ### Fixed